% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{A random dataset}
\format{
A dataset with 3 rows and 3 columns

\describe{
\item{jou}{arbitrary}
\item{nix}{also arbitrary}
\item{hey}{arbitrary again}

}
}
\usage{
mydata
}
\description{
A dataset containing some arbitrary values
}
\keyword{datasets}
