% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc1.R
\name{sub_add}
\alias{sub_add}
\title{Yeah}
\usage{
sub_add(x, y, z)
}
\arguments{
\item{x}{A numeric vector (or scalar)}

\item{y}{Lower limits of the distribution}

\item{z}{Upper limits of the distribution}
}
\value{
The output is a vector with the same length as \code{x}
}
\description{
\code{sub_add} adds 1 and subtracts 2 to values from a uniform distribution
}
\examples{
sub_add(1,2,4)

\dontrun{
sub_add("a")
}


}
